move code for FormOrderFragment from GlobalVars. Delete some unnecessary code and files
