--- app/src/main/java/com/amber/armtp/annotations/PGShowing.java
@@ -1,12 +1,0 @@
-package com.amber.armtp.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface PGShowing {
-    boolean isCanceled() default false;
-}
