+++ app/src/main/java/com/amber/armtp/interfaces/ServerChecker.java
@@ -1,17 +1,17 @@
 package com.amber.armtp.interfaces;
 
+import android.content.Context;
+
 import com.amber.armtp.Config;
 import com.amber.armtp.ServerDetails;
-import com.amber.armtp.annotations.PGShowing;
+import com.amber.armtp.extra.ProgressBarShower;
 import com.amber.armtp.ftp.Ping;
 
 public interface ServerChecker {
-    default void runCheckServerForAvailability(Thread t) {
+    default void runCheckServerForAvailability(Context context, Thread t) {
         try {
-            new Thread(new Runnable() {
-                @Override
-                @PGShowing
-                public void run() {
+            new Thread(() -> {
+                new ProgressBarShower(context).setFunction(() -> {
                     try {
                         if (!new Ping(ServerDetails.getInstance()).isReachable()) {
                             Config.sout("Сервер недоступен");
@@ -21,7 +21,8 @@ public interface ServerChecker {
                     } catch (Exception e) {
                         e.printStackTrace();
                     }
-                }
+                    return null;
+                }).start();
             }).start();
         } catch (Exception e) {
             Config.sout(e);
